+++ mantis-control-plane/mantis-control-plane-core/src/main/java/io/mantisrx/server/core/config/MantisExtension.java
@@ -14,18 +14,11 @@
  * limitations under the License.
  */
 
-package io.mantisrx.server.master.config;
+package io.mantisrx.server.core.config;
 
 import akka.actor.ActorSystem;
 import com.typesafe.config.Config;
-import lombok.val;
 
-public class MantisExtensionFactory {
-
-    @SuppressWarnings({"unchecked"})
-    public static <T> T createObject(Config config, ActorSystem actorSystem) throws Exception {
-        String className = config.getString("mantisExtensionClass");
-        val extension = (MantisExtension<T>) Class.forName(className).newInstance();
-        return extension.createObject(config, actorSystem);
-    }
+public interface MantisExtension<T> {
+    T createObject(Config config, ActorSystem actorSystem);
 }
