+++ mantis-control-plane/mantis-control-plane-core/src/main/java/io/mantisrx/server/core/config/MantisExtensionFactory.java
@@ -14,11 +14,19 @@
  * limitations under the License.
  */
 
-package io.mantisrx.server.master.config;
+package io.mantisrx.server.core.config;
 
 import akka.actor.ActorSystem;
 import com.typesafe.config.Config;
+import javax.annotation.Nullable;
+import lombok.val;
 
-public interface MantisExtension<T> {
-    T createObject(Config config, ActorSystem actorSystem);
+public class MantisExtensionFactory {
+
+    @SuppressWarnings({"unchecked"})
+    public static <T> T createObject(Config config, @Nullable ActorSystem actorSystem) throws Exception {
+        String className = config.getString("class");
+        val extension = (MantisExtension<T>) Class.forName(className).newInstance();
+        return extension.createObject(config, actorSystem);
+    }
 }
