+++ mantis-control-plane/mantis-control-plane-core/src/main/resources/reference.conf
@@ -1,6 +1,6 @@
 mantis {
     highAvailability {
-        mode = "local"
+        class = "io.mantisrx.server.core.zookeeper.ZookeeperHighAvailabilityServicesFactory"
         zookeeper {
             connectionTimeout = 10000 milliseconds
             connectionRetrySleepTime = 500 milliseconds
