+++ mantis-control-plane/mantis-control-plane-server/src/main/java/io/mantisrx/master/resourcecluster/resourceprovider/FileResourceClusterStorageProviderExtension.java
@@ -18,7 +18,7 @@ package io.mantisrx.master.resourcecluster.resourceprovider;
 
 import akka.actor.ActorSystem;
 import com.typesafe.config.Config;
-import io.mantisrx.server.master.config.MantisExtension;
+import io.mantisrx.server.core.config.MantisExtension;
 import java.io.File;
 
 @SuppressWarnings("unused")
