+++ mantis-control-plane/mantis-control-plane-server/src/main/resources/reference.conf
@@ -22,7 +22,7 @@ mantis {
             rulesetRefreshInterval = 3 minutes
         }
         store {
-            mantisExtensionClass = io.mantisrx.master.resourcecluster.resourceprovider.FileResourceClusterStorageProviderExtension
+            class = io.mantisrx.master.resourcecluster.resourceprovider.FileResourceClusterStorageProviderExtension
             directory = "/tmp/MantisSpool"
         }
     }
@@ -76,7 +76,7 @@ mantis {
         cache {
             maxArchivedJobs = 1000
         }
-        mantisExtensionClass = io.mantisrx.server.master.store.FileBasedStoreFactory
+        keyValueStore.class = io.mantisrx.server.master.store.FileBasedStoreFactory
     }
 
     initTimeout = 240 seconds
